+++ tests/app/routers/test_tools.py
@@ -78,18 +78,113 @@ async def test_get_tool_calls(
         assert tool_calls.json() == {"detail": {"detail": "Message not found."}}
 
         # Get the messages of the thread
-        messages = app_client.get(f"/threads/{thread_id}").json()
-        message_id = messages[-1]["message_id"]
+        messages = app_client.get(f"/threads/{thread_id}/messages").json()
+        message_id = messages["results"][-1]["message_id"]
         tool_calls = app_client.get(f"/tools/{thread_id}/{message_id}").json()
 
     assert (
-        tool_calls[0]
+        tool_calls["results"][0]
         == ToolCallSchema(
             tool_call_id="mock_tc_id",
             name="get_weather",
             arguments={"location": "Geneva"},
         ).model_dump()
     )
+
+
+@pytest.mark.httpx_mock(can_send_already_matched_responses=True)
+@pytest.mark.asyncio
+async def test_get_tool_calls_pagination(
+    patch_required_env,
+    httpx_mock,
+    app_client,
+    db_connection,
+    mock_openai_client,
+    get_weather_tool,
+):
+    routine = AgentsRoutine(client=mock_openai_client)
+
+    mock_openai_client.set_sequential_responses(
+        [
+            create_mock_response(
+                message={"role": "assistant", "content": ""},
+                function_calls=[
+                    {"name": "get_weather", "args": {"location": "Geneva"}}
+                ],
+            ),
+            create_mock_response(
+                message={"role": "assistant", "content": ""},
+                function_calls=[
+                    {"name": "get_weather", "args": {"location": "Montreux"}}
+                ],
+                id="mock_great_id",
+            ),
+            create_mock_response(
+                message={"role": "assistant", "content": ""},
+                function_calls=[
+                    {"name": "get_weather", "args": {"location": "Lausanne"}}
+                ],
+                id="mock_amazing_id",
+            ),
+            create_mock_response(
+                {"role": "assistant", "content": "sample response content"}
+            ),
+        ]
+    )
+    agent = Agent(tools=[get_weather_tool])
+
+    app.dependency_overrides[get_agents_routine] = lambda: routine
+    app.dependency_overrides[get_starting_agent] = lambda: agent
+    test_settings = Settings(
+        db={"prefix": db_connection},
+    )
+    app.dependency_overrides[get_settings] = lambda: test_settings
+    httpx_mock.add_response(
+        url=f"{test_settings.virtual_lab.get_project_url}/test_vlab/projects/test_project"
+    )
+
+    with app_client as app_client:
+        wrong_response = app_client.get("/tools/test/1234")
+        assert wrong_response.status_code == 404
+        assert wrong_response.json() == {"detail": {"detail": "Thread not found."}}
+
+        # Create a thread
+        create_output = app_client.post(
+            "/threads/?virtual_lab_id=test_vlab&project_id=test_project"
+        ).json()
+        thread_id = create_output["thread_id"]
+
+        # Fill the thread
+        app_client.post(
+            f"/qa/chat/{thread_id}",
+            json={"query": "This is my query"},
+            params={"thread_id": thread_id},
+            headers={"x-virtual-lab-id": "test_vlab", "x-project-id": "test_project"},
+        )
+
+        tool_calls = app_client.get(f"/tools/{thread_id}/wrong_id")
+        assert tool_calls.status_code == 404
+        assert tool_calls.json() == {"detail": {"detail": "Message not found."}}
+
+        # Get the messages of the thread
+        messages = app_client.get(f"/threads/{thread_id}/messages").json()
+        message_id = messages["results"][-1]["message_id"]
+        tool_calls_page_1 = app_client.get(
+            f"/tools/{thread_id}/{message_id}", params={"page_size": 2, "page": 1}
+        ).json()
+        tool_calls_page_2 = app_client.get(
+            f"/tools/{thread_id}/{message_id}", params={"page_size": 2, "page": 2}
+        ).json()
+
+    assert tool_calls_page_1["page"] == 1
+    assert tool_calls_page_1["page_size"] == 2
+    assert tool_calls_page_1["total_pages"] == 2
+    assert len(tool_calls_page_1["results"]) == 2
+
+    assert tool_calls_page_2["page"] == 2
+    assert tool_calls_page_2["page_size"] == 2
+    assert tool_calls_page_2["total_pages"] == 2
+    assert len(tool_calls_page_2["results"]) == 1
 
 
 @pytest.mark.httpx_mock(can_send_already_matched_responses=True)
@@ -153,11 +248,11 @@ async def test_get_tool_output(
         assert tool_output.json() == []
 
         # Get the messages of the thread
-        messages = app_client.get(f"/threads/{thread_id}").json()
-        message_id = messages[-1]["message_id"]
+        messages = app_client.get(f"/threads/{thread_id}/messages").json()
+        message_id = messages["results"][-1]["message_id"]
         tool_calls = app_client.get(f"/tools/{thread_id}/{message_id}").json()
 
-        tool_call_id = tool_calls[0]["tool_call_id"]
+        tool_call_id = tool_calls["results"][0]["tool_call_id"]
         tool_output = app_client.get(f"/tools/output/{thread_id}/{tool_call_id}")
 
     assert tool_output.json() == [json.dumps({"assistant": agent_2.name})]